# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_request.AuthLoginRequest:
    properties:
      email:
        description: 'Email    string `json:"email" binding:"required,email,max=255"`
          // TODO: re-enable email validation for production'
        maxLength: 255
        type: string
      password:
        description: 'Password string `json:"password" binding:"required,min=8,max=72"`
          // TODO: re-enable min=8 for production'
        maxLength: 72
        minLength: 3
        type: string
    required:
    - email
    - password
    type: object
  github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_request.AuthRegisterRequest:
    properties:
      email:
        maxLength: 255
        type: string
      full_name:
        maxLength: 255
        minLength: 3
        type: string
      password:
        maxLength: 72
        minLength: 8
        type: string
    required:
    - email
    - full_name
    - password
    type: object
...
  github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_request.ForgotPasswordRequest:
    properties:
      email:
        maxLength: 255
        type: string
    required:
    - email
    type: object
...
  github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_request.ResetPasswordRequest:
    properties:
      email:
        maxLength: 255
        type: string
      new_password:
        description: 'NewPassword string `json:"new_password" binding:"required,min=8,max=100"`
          // TODO: re-enable min=8 for production'
        maxLength: 100
        minLength: 3
        type: string
      otp:
        type: string
    required:
    - email
    - new_password
    - otp
    type: object
...
  github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response.FileResponse:
    properties:
      created_at:
        type: string
      file_name:
        type: string
      file_path:
        type: string
      file_size:
        type: integer
      file_url:
        type: string
      id:
        type: string
      mime_type:
        type: string
      type:
        $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_model.FileType'
      updated_at:
        type: string
      uploaded_by:
        type: string
    type: object
  github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response.LoginResponse:
    properties:
      access_token:
        type: string
      token_type:
        type: string
      user:
        $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response.UserData'
    type: object
...
  github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response.RegisterResponse:
    properties:
      id:
        type: string
    type: object
  github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response.UserData:
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      role:
        $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_model.UserRole'
      updated_at:
        type: string
    type: object
...
  github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.ActionSuccessResponse:
    properties:
      id:
        type: string
    type: object
...
  github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.ErrorCode:
    enum:
    - BAD_REQUEST
    - VALIDATION_ERROR
    - AUTH_REQUIRED
    - AUTH_INVALID_CREDENTIALS
    - NOT_FOUND
    - FORBIDDEN
    - INTERNAL_SERVER_ERROR
    - DATABASE_ERROR
    - TOO_MANY_REQUESTS
    type: string
    x-enum-varnames:
    - BAD_REQUEST
    - VALIDATION_ERROR
    - AUTH_REQUIRED
    - AUTH_INVALID_CREDENTIALS
    - NOT_FOUND
    - FORBIDDEN
    - INTERNAL_SERVER_ERROR
    - DATABASE_ERROR
    - TOO_MANY_REQUESTS
  github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.ErrorSchema:
    properties:
      code:
        $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.ErrorCode'
      details: {}
      message:
        type: string
    type: object
  github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.ErrorSchema'
      success:
        type: boolean
    type: object
...
  ? github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_FileResponse
  : properties:
      data:
        $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response.FileResponse'
      error:
        $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.ErrorSchema'
      success:
        type: boolean
    type: object
  ? github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_LoginResponse
  : properties:
      data:
        $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response.LoginResponse'
      error:
        $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.ErrorSchema'
      success:
        type: boolean
    type: object
...
  ? github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_RegisterResponse
  : properties:
      data:
        $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response.RegisterResponse'
      error:
        $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.ErrorSchema'
      success:
        type: boolean
    type: object
...
  ? github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common_ActionSuccessResponse
  : properties:
      data:
        $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.ActionSuccessResponse'
      error:
        $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.ErrorSchema'
      success:
        type: boolean
    type: object
...
  github_com_afifudin23_absensi-king-royal-api_internal_model.FileType:
    enum:
    - check_in
    - check_out
    - profile_picture
    - sick
    - extra_off
    - overtime
    - leave
    type: string
    x-enum-varnames:
    - FileTypeCheckIn
    - FileTypeCheckOut
    - FileTypeProfilePicture
    - FileTypeSick
    - FileTypeExtraOff
    - FileTypeOvertime
    - FileTypeLeave
...
  github_com_afifudin23_absensi-king-royal-api_internal_model.UserRole:
    enum:
    - admin
    - user
    type: string
    x-enum-varnames:
    - UserRoleAdmin
    - UserRoleUser
...
host: localhost:8080
info:
  contact:
    email: afifudin23@example.com
    name: Afifudin
  description: REST API untuk manajemen absensi, karyawan, dan penggajian King Royal.
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: Absensi King Royal API
  version: "1.0"
paths:
...
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: |-
        Meminta kode OTP yang dikirim ke email untuk proses reset password.

        **Request Body:**
        - `email` (string, required, format email): Alamat email akun yang terdaftar

        **Behavior:**
        - Jika email terdaftar, kode OTP (6 digit) akan dikirim ke alamat email tersebut
        - Jika email tidak terdaftar, response tetap 200 (untuk keamanan, tidak mengekspos data)
        - Kode OTP memiliki masa berlaku terbatas

        **Response 200:** Pesan konfirmasi pengiriman OTP
      parameters:
      - description: Email terdaftar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_request.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
      summary: Lupa password
      tags:
      - Auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: |-
        Login menggunakan email dan password, mengembalikan JWT access token.

        **Request Body:**
        - `email` (string, required): Alamat email akun
        - `password` (string, required): Password akun

        **Validasi & Behavior:**
        - Jika email tidak ditemukan atau password salah, request ditolak dengan 401
        - Token yang dikembalikan digunakan sebagai Bearer token di header Authorization

        **Response 200:** `access_token` (JWT), `token_type` (Bearer), data user (id, nama, email, role)

        **Error:** 401 jika kredensial salah
      parameters:
      - description: Kredensial login
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_request.AuthLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
      summary: Login pengguna
      tags:
      - Auth
...
  /auth/register:
    post:
      consumes:
      - application/json
      description: |-
        Mendaftarkan akun pengguna baru menggunakan email dan password.

        **Request Body:**
        - `full_name` (string, required, min 3 karakter): Nama lengkap pengguna
        - `email` (string, required, format email): Alamat email unik
        - `password` (string, required, min 8 karakter): Password akun

        **Validasi & Behavior:**
        - Email harus unik, jika sudah terdaftar akan mengembalikan 400
        - Password di-hash menggunakan argon2id sebelum disimpan

        **Response 201:** ID pengguna yang baru dibuat

        **Error:** 400 jika email sudah terdaftar atau validasi gagal
      parameters:
      - description: Data registrasi
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_request.AuthRegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
      summary: Registrasi pengguna baru
      tags:
      - Auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: |-
        Mereset password menggunakan kode OTP yang telah dikirim ke email.

        **Request Body:**
        - `email` (string, required, format email): Email akun yang akan direset
        - `otp` (string, required, tepat 6 karakter): Kode OTP dari email
        - `new_password` (string, required, min 8 karakter): Password baru

        **Validasi & Behavior:**
        - OTP harus valid dan belum kadaluarsa
        - Password baru akan di-hash sebelum disimpan
        - Setelah berhasil, OTP akan dihapus dan tidak bisa digunakan lagi

        **Response 200:** Pesan konfirmasi password berhasil direset

        **Error:** 400 jika OTP tidak valid atau sudah kadaluarsa
      parameters:
      - description: Data reset password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_request.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
      summary: Reset password
      tags:
      - Auth
  /files:
    post:
      consumes:
      - multipart/form-data
      description: |-
        Upload file ke server untuk digunakan pada check-in, check-out, foto profil, atau bukti pengajuan.

        **Form Data:**
        - `file` (file, required): File yang akan diupload. Maksimal 5 MB
        - `file_type` (string, required): Tipe file — `check_in`, `check_out`, `profile_picture`, `sick`, `extra_off`, `overtime`, `leave`

        **Validasi:**
        - Ukuran file maksimal 5 MB
        - Tipe file harus salah satu dari nilai yang diizinkan di atas
        - Jika tipe `profile_picture`, foto profil lama akan otomatis dihapus dan profil pengguna diperbarui

        **Response 200:** Data file yang berhasil diupload (id, url, nama, ukuran, mime type)

        **Cara pakai:** Upload file terlebih dahulu, lalu gunakan `id` yang dikembalikan di endpoint lain (check-in, pengajuan, dll)
      parameters:
      - description: File yang akan diupload (maks 5 MB)
        in: formData
        name: file
        required: true
        type: file
      - description: 'Tipe file: check_in | check_out | profile_picture | sick | extra_off
          | overtime | leave'
        in: formData
        name: file_type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_FileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
      security:
      - BearerAuth: []
      summary: Upload file
      tags:
      - Files
  /files/{file_id}:
    delete:
      description: |-
        Menghapus file dari server berdasarkan ID.

        **Path Param:**
        - `file_id` (UUID): ID file yang akan dihapus

        **Behavior:** File akan dihapus permanen dari server dan database

        **Response 200:** ID file yang berhasil dihapus
      parameters:
      - description: ID file
        in: path
        name: file_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common_ActionSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
      security:
      - BearerAuth: []
      summary: Hapus file
      tags:
      - Files
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token JWT dengan format: **Bearer &lt;token&gt;** — contoh:
      `Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...